refactor: Single test helpers directory
